Switch to uint4

+++ src/DX12GraphicsEngine/Engine/DXR/assets/shaders/Common_unbound.hlsl
@@ -38,8 +38,8 @@ cbuffer SceneShaderData : register(b2)
 	//meshes in the model ie the number of vertex buffers in the model.  For ex, if numberOfMeshes[0]=3,
 	//then model 0 has 3 unique mesh objects (ie 3 unique vertex buffers).
 
-	//CB arrays must have EACH element on 16 byte boundary, thus we are using float4
-	float4 numberOfMeshes[128];
+	//CB arrays must have EACH element on 16 byte boundary, thus we are using uint4
+	uint4 numberOfMeshes[128];
 };
 
 
@@ -50,8 +50,8 @@ cbuffer SceneTextureShaderData : register(b3)
 	//We then use that index into the actual texture array "diffuse_textures"  For ex, if texIndex=1, then the
 	//texture2D resource is accessed via diffuse_textures[1] ie diffuse_textures[texIndex].
 
-	//CB arrays must have EACH element on 16 byte boundary, thus we are using float4
-	float4 diffuseTextureIndexForMesh[128];
+	//CB arrays must have EACH element on 16 byte boundary, thus we are using uint4
+	uint4 diffuseTextureIndexForMesh[128];
 }
 
 // ---[ Resources ]---
